TaskManagement/src/main/java/ro/app/taskmanagement/validation/EmailValidator.java
package ro.app.taskmanagement.validation;

import ro.app.taskmanagement.exceptions.InvalidEmailException;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import java.util.regex.Pattern;

public class EmailValidator implements ConstraintValidator<ValidEmail, String> {
    private static final String EMAIL_PATTERN = "^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@" + "[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";

    @Override
    public void initialize(ValidEmail constraintAnnotation) {
    }

    @Override
    public boolean isValid(final String email, final ConstraintValidatorContext context) {

        if (!Pattern.compile(EMAIL_PATTERN).matcher(email).matches())
            throw new InvalidEmailException("Email is invalid!");

        return true;
    }
}
